+++ server/component/style/login/LoginController.php
@@ -29,39 +29,56 @@ class LoginController extends BaseController
     public function __construct($model)
     {
         parent::__construct($model);
-        if($model->is_logged_in()) $model->logout();
+        $this->login();
+    }
+
+    /* Private Methods *******************************************************/
+
+    /**
+     * Handle mobile device registration or redirect user
+     * 
+     * @param string $redirectUrl URL to redirect to for non-mobile requests
+     */
+    private function handleMobileOrRedirect($redirectUrl = null)
+    {
+        if (isset($_POST['mobile']) && $_POST['mobile']) {
+            // Set device id for the user
+            $device_token = isset($_POST['device_token']) ? $_POST['device_token'] : 'web';
+            $device_id = isset($_POST['device_id']) ? $_POST['device_id'] : 'web';
+            return $this->model->set_device_id_and_token($device_id, $device_token);
+        } else if ($redirectUrl) {
+            header('Location: ' . $redirectUrl);
+        }
+    }
+
+    /**
+     * Login the user
+     */
+    private function login() {
+        if($this->model->is_logged_in()) $this->model->logout();
 
         $this->failed = false;
 
-        if ($model->is_anonymous_users()) {
+        if ($this->model->is_anonymous_users()) {
             if (isset($_POST['type']) && $_POST['type'] == 'login' && isset($_POST['user_name']) && isset($_POST['password'])) {
-                if ($model->check_login_credentials_user_name($_POST['user_name'], $_POST['password'])) {
-                    if (isset($_POST['mobile']) && $_POST['mobile']) {
-                        // set device id for the user
-                        $device_token = isset($_POST['device_token']) ? $_POST['device_token'] : 'web';
-                        $device_id = isset($_POST['device_id']) ? $_POST['device_id'] : 'web';
-                        $res = $this->model->set_device_id_and_token($device_id, $device_token);
-                    } else {
-                        header('Location: ' . $model->get_target_url());
-                    }
-                } else
+                if ($this->model->check_login_credentials_user_name($_POST['user_name'], $_POST['password'])) {
+                    $this->handleMobileOrRedirect($this->model->get_target_url());
+                } else {
                     $this->failed = true;
+                }
             }
         } else {
             if (isset($_POST['type']) && $_POST['type'] == 'login' && isset($_POST['email']) && isset($_POST['password'])) {
-                if ($model->check_login_credentials($_POST['email'], $_POST['password'])) {
-                    if (isset($_POST['mobile']) && $_POST['mobile']) {
-                        // set device id for the user
-                        $device_token = isset($_POST['device_token']) ? $_POST['device_token'] : 'web';
-                        $device_id = isset($_POST['device_id']) ? $_POST['device_id'] : 'web';
-                        $res = $this->model->set_device_id_and_token($device_id, $device_token);
-                    } else {
-                        header('Location: ' . $model->get_target_url());
-                    }
-                } else
+                $res = $this->model->check_login_credentials($_POST['email'], $_POST['password']);
+                if ($res === '2fa') {
+                    $this->handleMobileOrRedirect($this->model->get_link_url(SH_TWO_FACTOR_AUTHENTICATION));
+                } else if ($res) {
+                    $this->handleMobileOrRedirect($this->model->get_target_url());
+                } else {
                     $this->failed = true;
+                }
             }
-        }
+        }   
     }
 
     /* Public Methods *********************************************************/