+++ server/component/style/login/tpl_login.php
@@ -21,6 +21,7 @@
             </div>
             <button type="submit" class="w-100 btn btn-<?php echo $this->type; ?>"><?php echo $this->login_label; ?></button>
         </form>
-        <a href="<?php echo $reset_url; ?>" class="small float-end text-<?php echo $this->type; ?>"><?php echo $this->reset_label; ?></a>
+        <a href="<?php echo $reset_url; ?>" class="small float-right text-<?php echo $this->type; ?>"><?php echo $this->reset_label; ?></a>
+        <?php $this->output_external_auth(); ?>
     </div>
 </div>