+++ server/page/BasePage.php
@@ -740,12 +740,8 @@ abstract class BasePage
      *  The unique identifier of the component.
      */
     public function output_component($key)
-<<<<<<< HEAD
     {        
-=======
-    {
         $this->services->get_clockwork()->startEvent('[BasePage][output_component] key:' .$key . " keyword: " . $this->keyword);
->>>>>>> main
         $component = $this->get_component($key);
         if ($component != null) {
             $component->output_content();